+++ src/containers/Prefs/PrefsRouted.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import  PrefsGeneral  from './PrefsGeneral';
+import  PrefsVars  from './PrefsVars';
 import  PrefsAdvanced  from './PrefsAdvanced';
 
 export class PrefsRouted extends React.Component {
@@ -10,21 +11,28 @@ export class PrefsRouted extends React.Component {
       <Switch>
         <Route
           path={'/prefs/general'}
-          render={({history,match})=>{
+          render={()=>{
             return (<PrefsGeneral
             />);
           }}
         />
         <Route
+          path={'/prefs/vars'}
+          render={()=>{
+            return (<PrefsVars
+            />);
+          }}
+        />
+        <Route
           path={'/prefs/advanced'}
-          render={({history,match})=>{
+          render={()=>{
             return (<PrefsAdvanced
             />);
           }}
         />
         <Route
           path={'/prefs'}
-          render={({history,match})=>{
+          render={()=>{
             return (<PrefsGeneral
             />);
           }}