+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Quiqr App CHANGELOG
+**WORK IN PROGRESS**
+- fix: improve first start
+
 ## 0.19.5 (2025-02-20)
 - fix: whitescreen in compiled version
 - feature: add txtInsertButtons option to string fieldtype [documentation](https://book.quiqr.org/docs/20-quiqr-developer-reference/03-content-model/03-form-fields/data-field-types/string/)